feat: finish signup refactor first step(some files need to be deleted)
