+++ packages/rn-sdk/js/config/entries.ts
@@ -3,7 +3,6 @@ export enum PortkeyEntries {
 
   SIGN_IN_ENTRY = 'sign_in_entry',
   SIGN_UP_ENTRY = 'sign_up_entry',
-  SIGN_UP_REFERRAL_ENTRY = 'sign_up_referral_entry',
   SELECT_COUNTRY_ENTRY = 'select_country_entry',
   GUARDIAN_APPROVAL_ENTRY = 'guardian_approval_entry',
   VERIFIER_DETAIL_ENTRY = 'verifier_detail_entry',
